Refactore GUI into multiple files
